Numeric.LBFGSB.Convenience

Description

Some functions that can be useful together with L-BFGS-B.

Documentation

approximateGradient :: Vector v Double => Double -> (v Double -> Double) -> v Double -> v Double

approximateGradient h f x is an approximation of the gradient of f at x, computed using central differences with step size h.

listFunction :: ([Double] -> Double) -> Vector Double -> Double

Turn a function on lists into one on Storable Vectors.

vectorFunction :: Vector v Double => (v Double -> Double) -> Vector Double -> Double

Turn a function on any generic Vectors into one on Storable Vectors.