Numeric.LBFGSB.Result

Description

The Result data type encodes the minimization solution, as well as auxiliary information about the minimization process.

Documentation

data Result

Stores the result of the minimization process.

Constructors

Result

Fields

solution :: Vector Double

Solution point if the minimization completed successfully. See stopReason.

backtrace :: [Vector Double]

The steps taken to reach the solution, in reverse order. Does not include the starting point.

stopReason :: StopReason

The reason L-BFGS-B terminated. Only if this is Converged should you consider the solution correct!

data StopReason

The reason L-BFGS-B terminated.

Constructors

Converged
The solution converged.

StepCount
The number of steps exceeded the user's request.

Other String
Something else occured. In Other s, s is the contents of L-BFGS-B's task variable on exit, as documented in the source code of L-BFGS-B itself.